Move observer and test in directory structure
